pyproject.toml
[tool.poetry]
name = "aracor-ai"
version = "0.1.0"
description = "An example package demonstrating dependency management and code quality setup."
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "aracor_ai" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
langchain = "^0.3.19"
langchain-openai = "^0.3.6"
python-dotenv = "^1.0.1"
langchain-cohere = "^0.4.2"
pdfplumber = "^0.11.5"
python-docx = "^1.1.2"

[tool.poetry.group.development.dependencies]
pytest = "^8.2"
black = "^25.1"
isort = "^6.0"
mypy = "^1.5"
pylint = "^3.3.4"
# Add any other development tools as needed


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.25.3"
pylint = "^3.3.4"
black = "^25.1.0"
isort = "^6.0.0"

[tool.black]
line-length = 130
target-version = ["py311", "py312"]

[tool.isort]
profile = "black"
line_length = 130

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
